+++ src/main/java/zabbum/oelremake/ChangePrices.java
@@ -5,6 +5,7 @@ import com.googlecode.lanterna.TextColor;
 import com.googlecode.lanterna.graphics.SimpleTheme;
 import com.googlecode.lanterna.gui2.*;
 import com.googlecode.lanterna.gui2.table.Table;
+import zabbum.oelremake.Plants.Industries.AbstractIndustry;
 
 import java.util.regex.Pattern;
