+++ src/main/java/zabbum/oelremake/Plants/AbstractPlant.java
@@ -1,7 +1,8 @@
-package zabbum.oelremake;
+package zabbum.oelremake.Plants;
 
 import lombok.Getter;
 import lombok.Setter;
+import zabbum.oelremake.Player;
 
 @Getter
 @Setter
