+++ src/main/java/zabbum/oelremake/Plants/Industries/Cars/CarsOperations.java
@@ -1,4 +1,4 @@
-package zabbum.oelremake.Cars;
+package zabbum.oelremake.Plants.Industries.Cars;
 
 import com.googlecode.lanterna.TextColor;
 import com.googlecode.lanterna.gui2.Window;
