+++ src/main/java/zabbum/oelremake/Plants/Industries/Pumps/PumpsOperations.java
@@ -1,4 +1,4 @@
-package zabbum.oelremake.Drills;
+package zabbum.oelremake.Plants.Industries.Pumps;
 
 import com.googlecode.lanterna.TextColor;
 import com.googlecode.lanterna.gui2.Window;
@@ -12,28 +12,26 @@ import zabbum.oelremake.Player;
 import java.util.Map;
 
 @Getter
-public class DrillsOperations implements BuyableIndustryOperation, BuyableProductsOperation, SabotableOperation {
+public class PumpsOperations implements BuyableIndustryOperation, BuyableProductsOperation, SabotableOperation {
 
-    private final DrillsIndustry[] industries;
+    private final PumpsIndustry[] industries;
 
-    public DrillsOperations() { industries = initialize(); }
+    public PumpsOperations() { industries = initialize(); }
 
     /**
      * Initialize drills industries.
      *
      * @return array of cars industries
      */
-    public static DrillsIndustry[] initialize() {
-        DrillsIndustry[] drillProds = new DrillsIndustry[3];
+    public static PumpsIndustry[] initialize() {
+        PumpsIndustry[] pumpProds = new PumpsIndustry[2];
 
-        // Drills productions initialization
-        drillProds[0] = new DrillsIndustry("TURBOWIERT");
-        drillProds[1] = new DrillsIndustry("NA B£YSK INC.");
-        drillProds[2] = new DrillsIndustry("PET SHOP&BOYS");
+        // Pump productions initialization
+        pumpProds[0] = new PumpsIndustry("ZASSANICKI GMBH");
+        pumpProds[1] = new PumpsIndustry("DR PUMPENER");
 
-        return drillProds;
+        return pumpProds;
     }
-
 
     @Override
     public void buyIndustryMenu(Player player, Window window, Map<String, String> langMap) throws InterruptedException {
@@ -47,10 +45,10 @@ public class DrillsOperations implements BuyableIndustryOperation, BuyableProduc
                 TextColor.ANSI.WHITE_BRIGHT,
                 TextColor.ANSI.CYAN,
                 TextColor.ANSI.BLUE,
-                langMap.get("drillsIndustrySale"),
-                langMap.get("drillsIndustryPrompt"),
-                langMap.get("drillsPricePrompt"),
-                60000,
+                langMap.get("pumpsIndustrySale"),
+                langMap.get("pumpsIndustryPrompt"),
+                langMap.get("pumpsPricePrompt"),
+                50000,
                 window,
                 langMap
         );
@@ -58,22 +56,21 @@ public class DrillsOperations implements BuyableIndustryOperation, BuyableProduc
 
     @Override
     public void buyProductsMenu(Player player, Oilfield[] oilfields, Window window, Map<String, String> langMap) throws InterruptedException {
-
         BuyableProductsOperation.defaultBuyProductsMenu(
                 player,
                 industries,
                 oilfields,
                 TextColor.ANSI.WHITE_BRIGHT,
-                TextColor.ANSI.YELLOW,
-                TextColor.ANSI.BLUE,
+                TextColor.ANSI.BLACK_BRIGHT,
+                TextColor.ANSI.BLACK_BRIGHT,
                 TextColor.ANSI.WHITE_BRIGHT,
                 TextColor.ANSI.WHITE_BRIGHT,
                 TextColor.ANSI.CYAN,
-                TextColor.ANSI.YELLOW,
-                TextColor.ANSI.BLUE,
-                langMap.get("drillsHereYouCanBuy"),
-                langMap.get("drillsProductsAmountPrompt"),
-                10,
+                TextColor.ANSI.BLACK_BRIGHT,
+                TextColor.ANSI.BLACK,
+                langMap.get("pumpsHereYouCanBuy"),
+                langMap.get("pumpsProductsAmountPrompt"),
+                15,
                 window,
                 langMap
         );
